+++ src/django_bootstrap5/forms.py
@@ -1,10 +1,5 @@
-from django.utils.safestring import mark_safe
-
 from .core import get_field_renderer, get_form_renderer, get_formset_renderer
-from .css import merge_css_classes
-from .exceptions import BootstrapError
 from .html import render_tag
-from .size import DEFAULT_SIZE, SIZE_MD, get_size_class
 
 WRAPPER_CLASS = ""
 WRAPPER_TAG = "div"
@@ -50,58 +45,3 @@ def render_label(content, label_for=None, label_class=None, label_title=""):
     if label_title:
         attrs["title"] = label_title
     return render_tag("label", attrs=attrs, content=content)
-
-
-def render_button(
-    content,
-    button_type=None,
-    button_class="btn-primary",
-    button_outline=False,
-    size="",
-    href="",
-    name=None,
-    value=None,
-    title=None,
-    extra_classes="",
-    id="",
-):
-    """Render a button with content."""
-    attrs = {}
-    size_class = get_size_class(size, prefix="btn", skip=SIZE_MD, default=DEFAULT_SIZE)
-    classes = merge_css_classes("btn", button_class, size_class)
-
-    if button_type:
-        if button_type not in ("submit", "reset", "button", "link"):
-            raise BootstrapError(
-                (
-                    'Parameter "button_type" should be "submit", "reset", "button", "link" or empty '
-                    '("{button_type}" given).'
-                ).format(button_type=button_type)
-            )
-        if button_type != "link":
-            attrs["type"] = button_type
-
-    classes = merge_css_classes(classes, extra_classes)
-    attrs["class"] = classes
-
-    if href:
-        tag = "a"
-        if button_type and button_type != "link":
-            raise BootstrapError(
-                'Button of type "{button_type}" is not allowed a "href" parameter.'.format(button_type=button_type)
-            )
-        attrs["href"] = href
-        # Specify role for link with button appearance
-        attrs.setdefault("role", "button")
-    else:
-        tag = "button"
-
-    if id:
-        attrs["id"] = id
-    if name:
-        attrs["name"] = name
-    if value:
-        attrs["value"] = value
-    if title:
-        attrs["title"] = title
-    return render_tag(tag, attrs=attrs, content=mark_safe(content))
